# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "http server api docs",
        "title": "lol api docs",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/api/v1/loan": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "submit information to create loan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan"
                ],
                "summary": "create loan",
                "parameters": [
                    {
                        "description": "loan information",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateLoanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.CreateLoanReply"
                        }
                    }
                }
            }
        },
        "/api/v1/loan/list": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "list of loans by paging and conditions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan"
                ],
                "summary": "list of loans by query parameters",
                "parameters": [
                    {
                        "description": "query parameters",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Params"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ListLoansReply"
                        }
                    }
                }
            }
        },
        "/api/v1/loan/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get loan detail by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan"
                ],
                "summary": "get loan detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetLoanByIDReply"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update loan information by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan"
                ],
                "summary": "update loan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "loan information",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateLoanByIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UpdateLoanByIDReply"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "delete loan by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan"
                ],
                "summary": "delete loan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.DeleteLoanByIDReply"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "types.Column": {
            "type": "object",
            "properties": {
                "exp": {
                    "description": "expressions, which default to = when the value is null, have =, !=, \u003e, \u003e=, \u003c, \u003c=, like",
                    "type": "string"
                },
                "logic": {
                    "description": "logical type, default value is \"and\", support \u0026, and, ||, or",
                    "type": "string"
                },
                "name": {
                    "description": "column name",
                    "type": "string"
                },
                "value": {
                    "description": "column value"
                }
            }
        },
        "types.CreateLoanReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "id": {
                            "description": "id",
                            "type": "integer"
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
        "types.CreateLoanRequest": {
            "type": "object",
            "properties": {
                "carModel": {
                    "description": "车型",
                    "type": "string"
                },
                "carPlate": {
                    "description": "车牌",
                    "type": "string"
                },
                "createAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "loanMoney": {
                    "description": "借款金额",
                    "type": "number"
                },
                "loanPeriod": {
                    "description": "借款期数",
                    "type": "integer"
                },
                "loanReturnDate": {
                    "description": "还款日期",
                    "type": "string"
                },
                "mobile": {
                    "description": "手机号码",
                    "type": "string"
                },
                "monthlyPayment": {
                    "description": "每月应还",
                    "type": "integer"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "string"
                },
                "userID": {
                    "description": "身份证号码",
                    "type": "string"
                }
            }
        },
...
        "types.DeleteLoanByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data"
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.GetLoanByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "loan": {
                            "$ref": "#/definitions/types.LoanObjDetail"
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.ListLoansReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "loans": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.LoanObjDetail"
                            }
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.LoanObjDetail": {
            "type": "object",
            "properties": {
                "carModel": {
                    "description": "车型",
                    "type": "string"
                },
                "carPlate": {
                    "description": "车牌",
                    "type": "string"
                },
                "createAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "convert to uint64 id",
                    "type": "integer"
                },
                "loanMoney": {
                    "description": "借款金额",
                    "type": "number"
                },
                "loanPeriod": {
                    "description": "借款期数",
                    "type": "integer"
                },
                "loanReturnDate": {
                    "description": "还款日期",
                    "type": "string"
                },
                "mobile": {
                    "description": "手机号码",
                    "type": "string"
                },
                "monthlyPayment": {
                    "description": "每月应还",
                    "type": "integer"
                },
                "name": {
                    "description": "序号",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "string"
                },
                "userID": {
                    "description": "身份证号码",
                    "type": "string"
                }
            }
        },
        "types.Params": {
            "type": "object",
            "properties": {
                "columns": {
                    "description": "query conditions",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Column"
                    }
                },
                "limit": {
                    "description": "lines per page",
                    "type": "integer"
                },
                "page": {
                    "description": "page number, starting from page 0",
                    "type": "integer"
                },
                "sort": {
                    "description": "sorted fields, multi-column sorting separated by commas",
                    "type": "string"
                }
            }
        },
...
        "types.UpdateLoanByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data"
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
        "types.UpdateLoanByIDRequest": {
            "type": "object",
            "properties": {
                "carModel": {
                    "description": "车型",
                    "type": "string"
                },
                "carPlate": {
                    "description": "车牌",
                    "type": "string"
                },
                "createAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "uint64 id",
                    "type": "integer"
                },
                "loanMoney": {
                    "description": "借款金额",
                    "type": "number"
                },
                "loanPeriod": {
                    "description": "借款期数",
                    "type": "integer"
                },
                "loanReturnDate": {
                    "description": "还款日期",
                    "type": "string"
                },
                "mobile": {
                    "description": "手机号码",
                    "type": "string"
                },
                "monthlyPayment": {
                    "description": "每月应还",
                    "type": "integer"
                },
                "name": {
                    "description": "序号",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "string"
                },
                "userID": {
                    "description": "身份证号码",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type Bearer your-jwt-token to Value",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}